FastAPI/requirements.txt
fastapi
uvicorn[standard]
psycopg2-binary
python-multipart
passlib[bcrypt]
python-jose
pydantic[email]
python-dotenv
